requirements.txt
tensorflow-gpu>=1.15.5
keras>=2.3.1
numpy>=1.18.5
opencv-python>=4.5.3.56
tqdm>=4.62.2
h5py>=2.10.0
scipy>=1.5.4
scikit-image>=0.17.2
